Features
This node receives low level control messages that include the duty
cycle to which all rotors should be spin, and effectively sends the
desired commands to the PCA9685.

This node is used when the main control algorithms are being run outboard,
in another PC, and the onboard computer simply receives the desired
duty cycles and generates the appropriate signal.